# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{ws}/databases": {
      "get": {
        "operationId": "DatabasesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List databases (guests: scoped spaces only)",
        "tags": [
          "databases"
        ]
      },
      "post": {
        "operationId": "DatabasesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a database (auto: title field, system fields, default view)",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}": {
      "get": {
        "operationId": "DatabasesController_get",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Database with live fields and views (schema introspection)",
        "tags": [
          "databases"
        ]
      },
      "patch": {
        "operationId": "DatabasesController_update",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatabaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename / re-icon (creator); moving between spaces stays member+",
        "tags": [
          "databases"
        ]
      },
      "delete": {
        "operationId": "DatabasesController_remove",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDatabaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hard delete — body.confirm must equal the database name",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/fields": {
      "post": {
        "operationId": "FieldsController_create",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFieldDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a field (select types accept initial options)",
        "tags": [
          "fields"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/fields/{field}": {
      "patch": {
        "operationId": "FieldsController_update",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFieldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename / reconfigure / reorder a field",
        "tags": [
          "fields"
        ]
      },
      "delete": {
        "operationId": "FieldsController_remove",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a field (returns records_with_value)",
        "tags": [
          "fields"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/fields/{field}/usage": {
      "get": {
        "operationId": "FieldsController_usage",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "How many live records carry a value for this field",
        "tags": [
          "fields"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/fields/{field}/change-type": {
      "post": {
        "operationId": "FieldsController_changeType",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeFieldTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Convert field type within the compatibility matrix (dry_run supported)",
        "tags": [
          "fields"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/fields/{field}/options": {
      "post": {
        "operationId": "FieldsController_addOption",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a select option",
        "tags": [
          "fields"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/fields/{field}/options/{option}": {
      "patch": {
        "operationId": "FieldsController_updateOption",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "option",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename / recolor / reorder an option (O(1), ids are stable)",
        "tags": [
          "fields"
        ]
      },
      "delete": {
        "operationId": "FieldsController_removeOption",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "option",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an option — 409 with usage count unless confirm: true",
        "tags": [
          "fields"
        ]
      }
    },
...
    "/api/v1/workspaces/{ws}/databases/{db}/records/{rec}/document": {
      "get": {
        "operationId": "DocumentsController_get",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record description (BlockNote JSON; version 0 = never written)",
        "tags": [
          "documents"
        ]
      },
      "put": {
        "operationId": "DocumentsController_put",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Write the description — 409 with current version on conflict",
        "tags": [
          "documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "StoryOS API",
    "description": "The public StoryOS REST API. The web app is client #1 of this same API — everything the UI does, you can do.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDatabaseDto": {
        "type": "object",
        "properties": {
          "space_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "icon": {
            "type": "string",
            "maxLength": 16
          }
        },
        "required": [
          "space_id",
          "name"
        ]
      },
      "UpdateDatabaseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "icon": {
            "type": "string",
            "maxLength": 16,
            "nullable": true
          },
          "space_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "position": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        }
      },
      "DeleteDatabaseDto": {
        "type": "object",
        "properties": {
          "confirm": {
            "type": "string"
          },
          "sever_relations": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": [
          "confirm"
        ]
      },
      "CreateFieldDto": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number",
              "checkbox",
              "date",
              "select",
              "multi_select",
              "url",
              "email",
              "user"
            ]
          },
          "config": {
            "type": "object",
            "additionalProperties": {}
          },
          "options": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                },
                "color": {
                  "type": "string"
                }
              },
              "required": [
                "label"
              ]
            }
          }
        },
        "required": [
          "display_name",
          "type"
        ]
      },
      "UpdateFieldDto": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "config": {
            "type": "object",
            "additionalProperties": {}
          },
          "position": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        }
      },
      "ChangeFieldTypeDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number",
              "checkbox",
              "date",
              "select",
              "multi_select",
              "url",
              "email",
              "user"
            ]
          },
          "dry_run": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": [
          "type"
        ]
      },
      "CreateOptionDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "color": {
            "default": "gray",
            "type": "string",
            "enum": [
              "gray",
              "brown",
              "gold",
              "orange",
              "red",
              "pink",
              "purple",
              "blue",
              "teal",
              "green"
            ]
          }
        },
        "required": [
          "label"
        ]
      },
      "UpdateOptionDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "color": {
            "type": "string",
            "enum": [
              "gray",
              "brown",
              "gold",
              "orange",
              "red",
              "pink",
              "purple",
              "blue",
              "teal",
              "green"
            ]
          },
          "position": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        }
      },
      "DeleteOptionDto": {
        "type": "object",
        "properties": {
          "confirm": {
            "default": false,
            "type": "boolean"
          },
          "reassign_to": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        }
      },
...
      "PutDocumentDto": {
        "type": "object",
        "properties": {
          "content": {},
          "expected_version": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "content",
          "expected_version"
        ]
      },
...
    }
  }
}